+++ docs/CHANGELOG.md
@@ -4,12 +4,12 @@
 
 ### Breaking
 
-* All log argument is removed use `KIME_LOG` environment for set log level it can be `ERROR`, `WARN`, `INFO`, `DEBUG`, `TRACE` other variables equal to empty which means `OFF`
 * Now frontends check engine's version if it failed, must cause hard error
 * Use xim async only it will break some apps like neovide but fix many wrong bevaivors and performance
 
 ### Improve
 
+* New tool `kime-check` for diagnostic kime [#270](https://github.com/Riey/kime/issues/270)
 * Correct invalid sebolsik-390 builtin layout [#261](https://github.com/Riey/kime/issues/261)
 * Fix indicator can't load icon data [#260](https://github.com/Riey/kime/issues/260)
 * Fix choseong compose bug [#263](https://github.com/Riey/kime/issues/263)